+++ www/core/Protocols/OAuth/OAuthModule.php
@@ -27,7 +27,7 @@ use SimpleID\Auth\AuthManager;
 use SimpleID\Module;
 use SimpleID\ModuleManager;
 use SimpleID\Base\ScopeInfoCollectionEvent;
-use SimpleID\Models\ConsentEvent;
+use SimpleID\Base\ConsentEvent;
 use SimpleID\Protocols\ProtocolResult;
 use SimpleID\Store\StoreManager;
 use SimpleID\Util\SecurityToken;
